pyproject.toml
[build-system]
requires = ["hatchling>=1.20"]
build-backend = "hatchling.build"

[project]
name = "eventdbx"
dynamic = ["version"]
description = "Official (preview) Python client for the EventDBX ingestion API."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
  { name = "Patrick Thach", email = "thachp@gmail.com" }
]
keywords = ["eventdbx", "events", "analytics", "client", "sdk"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
  "httpx>=0.27,<0.29",
  "noiseprotocol>=0.3,<0.4",
  "pycapnp>=1.3,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5",
  "build>=1.2",
  "twine>=5.1",
]

[project.urls]
Homepage = "https://github.com/eventdbx/eventdbx-python"
Repository = "https://github.com/eventdbx/eventdbx-python"
"Bug Tracker" = "https://github.com/eventdbx/eventdbx-python/issues"
Documentation = "https://docs.eventdbx.com"

[tool.hatch.version]
path = "src/eventdbx/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/eventdbx"]

[tool.hatch.build.targets.wheel.force-include]
"src/eventdbx/proto/control.capnp" = "eventdbx/proto/control.capnp"

[tool.hatch.build.targets.sdist]
include = [
  "src/eventdbx",
  "tests",
  "README.md",
  "LICENSE",
]

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
pythonpath = ["src"]
